chore(frontend): refactor app router and providers
